fixed few functions

+++ src/app/api/usersApi.ts
@@ -2,73 +2,6 @@
 import axios from "axios";
 import { cookies } from "next/headers";
 
-const cookieStore = cookies();
-const token = cookieStore.get("authToken")?.value;
-if (!token) {
-  throw new Error("No auth token found");
-}
-
-interface User {
-  username: string;
-  password: string;
-  email?: string;
-  fullname?: string;
-}
-
-// SEARCH FOR FOLLOWING
-export const searchFollowing = async (query: string) => {
-  try {
-    const url = `http://localhost:4444/users/search/following?query=${query}`;
-    const { data } = await axios.get(url, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    return data;
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-// GET FOLLOWINGS USER
-export const getFollowingsUsers = async () => {
-  try {
-    const res = await axios.get("http://localhost:4444/users/following", {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    return res.data;
-  } catch (error: any) {
-    if (axios.isAxiosError(error)) {
-      throw new Error(error.response?.data.msg || "Unexpected error occurred");
-    }
-    throw new Error("Unexpected error occurred");
-  }
-};
-
-// SEARCH FOR USER
-export const searchUser = async (query: string) => {
-  try {
-    const url = `http://localhost:4444/users/search?query=${query}`;
-    const { data } = await axios.get(url, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    return data;
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-// GET USER INFO BY ID
-export const getUserByUsername = async (uname: any) => {
-  let res = await axios.get(`http://localhost:4444/users/${uname}`);
-  return res.data;
-};
-
 // REGISTER
 export const registerUser = async (user: User) => {
   try {
@@ -95,8 +28,93 @@ export const loginUser = async (user: User) => {
   }
 };
 
+interface User {
+  username: string;
+  password: string;
+  email?: string;
+  fullname?: string;
+}
+
+// SEARCH FOR FOLLOWING
+export const searchFollowing = async (query: string) => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
+  try {
+    const url = `http://localhost:4444/users/search/following?query=${query}`;
+    const { data } = await axios.get(url, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    return data;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+// GET FOLLOWINGS USER
+export const getFollowingsUsers = async () => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
+  try {
+    const res = await axios.get("http://localhost:4444/users/following", {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+
+    return res.data;
+  } catch (error: any) {
+    if (axios.isAxiosError(error)) {
+      throw new Error(error.response?.data.msg || "Unexpected error occurred");
+    }
+    throw new Error("Unexpected error occurred");
+  }
+};
+
+// SEARCH FOR USER
+export const searchUser = async (query: string) => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
+  try {
+    const url = `http://localhost:4444/users/search?query=${query}`;
+    const { data } = await axios.get(url, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    return data;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+// GET USER INFO BY ID
+export const getUserByUsername = async (uname: any) => {
+  let res = await axios.get(`http://localhost:4444/users/${uname}`);
+  return res.data;
+};
+
 // GET OWN USER INFO
 export const getUserInfo = async () => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
   const res = await axios.get("http://localhost:4444/users/owner", {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -107,6 +125,12 @@ export const getUserInfo = async () => {
 
 // GET ALL USERS
 export const getAllUsers = async () => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
   try {
     const res = await axios.get("http://localhost:4444/users/all", {
       headers: {
@@ -170,6 +194,12 @@ export const removePfp = async () => {
 
 // FOLLOW / UNFOL
 export const followUnfol = async (id: string) => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
   const res = await axios.post(
     `http://localhost:4444/follows/${id}`,
     {},
@@ -184,6 +214,12 @@ export const followUnfol = async (id: string) => {
 
 // DELETE FOLLOWER
 export const removeFoll = async (id: string) => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
   const res = await axios.delete(`http://localhost:4444/follows/${id}`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -194,6 +230,12 @@ export const removeFoll = async (id: string) => {
 
 // ADD ACC TO FAVOURTIE
 export const favUnfav = async (id: string) => {
+  const cookieStore = cookies();
+  const token = cookieStore.get("authToken")?.value;
+  if (!token) {
+    throw new Error("No auth token found");
+  }
+
   const res = await axios.post(
     `http://localhost:4444/favs/${id}`,
     {},
